pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ticktick-cli"
version = "0.1.0"
description = "CLI for reading and managing TickTick tasks, with Claude cowork integration"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ticktick = "ticktick.cli:main"
